+++ contrib/tlstest/client_12.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-../../cmd/dgraph-live-loader/dgraph-live-loader -d server1.dgraph.io:9080 -tls.on -tls.ca_certs ca.crt -tls.cert client.crt -tls.cert_key client.key -tls.server_name server1.dgraph.io -tls.min_version=TLS12 -r data.rdf.gz 
+../../dgraph/dgraph live -d server1.dgraph.io:9080 --tls.on --tls.ca_certs ca.crt --tls.cert client.crt --tls.cert_key client.key --tls.server_name server1.dgraph.io --tls.min_version=TLS12 -r data.rdf.gz -z 127.0.0.1:7081