+++ contrib/tlstest/run.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-# We run the TLS tests only when a commit is made on master/release
- if [[ $TRAVIS_BRANCH =~ master|release\/ ]] ; then
-	make test
- else
- 	echo "TLS tests skipped."
- fi
+pushd $GOPATH/src/github.com/dgraph-io/dgraph/contrib/tlstest
+make test
+popd