+++ contrib/tlstest/server_pass.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-../../cmd/dgraph/dgraph -tls.on -tls.ca_certs ca.crt -tls.cert server_pass.crt -tls.cert_key server_pass.key -tls.cert_key_passphrase secret --memory_mb 2048 --zero 127.0.0.1:8888
+../../dgraph/dgraph server --tls.on --tls.ca_certs ca.crt --tls.cert server_pass.crt --tls.cert_key server_pass.key --tls.cert_key_passphrase secret --memory_mb 2048 --zero 127.0.0.1:7081 &> dgraph.log