+++ contrib/tlstest/test.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
-killall dgraph dgraphzero > /dev/null 2>&1
+killall -9 dgraph
 
-DGRAPH_ROOT=$GOPATH/src/github.com/dgraph-io/dgraph/cmd
+DGRAPH_ROOT=$GOPATH/src/github.com/dgraph-io/dgraph/dgraph
 function build {
-  pushd $DGRAPH_ROOT/$1 > /dev/null
+  pushd $DGRAPH_ROOT > /dev/null
   go build .
   popd > /dev/null
 }
@@ -13,11 +13,9 @@ SERVER=$1
 CLIENT=$2
 EXPECTED=$3
 
-build "dgraphzero"
 build "dgraph"
-build "dgraph-live-loader"
 
-$DGRAPH_ROOT/dgraphzero/dgraphzero -w zw > zero.log 2>&1 &
+$DGRAPH_ROOT/dgraph zero -w zw -o 1 > zero.log 2>&1 &
 sleep 5