+++ x/tls_helper.go
@@ -59,8 +59,6 @@ func SetTLSFlags(conf *TLSHelperConfig, flag *pflag.FlagSet) {
 	flag.StringVar(&conf.Cert, "tls.cert", "", "Certificate file path.")
 	flag.StringVar(&conf.Key, "tls.cert_key", "", "Certificate key file path.")
 	flag.StringVar(&conf.KeyPassphrase, "tls.cert_key_passphrase", "", "Certificate key passphrase.")
-	flag.StringVar(&conf.ClientAuth, "tls.client_auth", "", "Enable TLS client authentication")
-	flag.StringVar(&conf.ClientCACerts, "tls.ca_certs", "", "CA Certs file path.")
 	flag.BoolVar(&conf.UseSystemClientCACerts, "tls.use_system_ca", false, "Include System CA into CA Certs.")
 	flag.StringVar(&conf.MinVersion, "tls.min_version", "TLS11", "TLS min version.")
 	flag.StringVar(&conf.MaxVersion, "tls.max_version", "TLS12", "TLS max version.")